requirements.txt

python-telegram-bot
python-telegram-bot[job-queue]
gspread
oauth2client
python-dotenv